# apps/api/src/docs/openapi.yaml
openapi: 3.0.3
info:
  title: ご近所サポートちょこっと API設計書
  version: 1.0.0
  description: |
    高齢者 × ボランティア マッチングアプリ「ちょこっと」のAPI仕様書です。  
    各エンドポイントの概要・入力・出力・ステータスコードを示します。
servers:
  - url: http://localhost:3001/api
    description: 開発環境
...
paths:
...
  /user/requests:
    get:
      tags:
        - User
      summary: 利用者自身が依頼した依頼の一覧を取得
      description: |
        ログイン中の利用者が作成した依頼のうち、  
        status が completed / canceled / expired 以外の依頼を返します。
      security:
        - bearerAuth: []
      responses:
        "200":
          description: 依頼一覧取得成功
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/UserRequest"
        "403":
          description: 利用者のみ依頼可能です。
        "500":
          description: サーバーエラー

    post:
      tags:
        - User
      summary: 新規依頼を登録
      description: |
        ログイン中の利用者が新しい依頼を登録します。  
        バリデーションは Zod スキーマ `RequestCreateSchema` によって行われます。  
        title は固定で「買い物代行」が設定されます。
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/NewRequestInput"              
      responses:
        "201":
          description: 新規依頼登録成功
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/NewRequest"
        "400":
          description: Zod バリデーションエラー
        "403":
          description: 利用者のみ依頼可能です。
        "500":
          description: サーバーエラー
  /user/requests/{requestId}:
    get:
      tags:
        - User
      summary: 利用者向けの依頼詳細を取得
      description: |
        ログイン中の利用者が自分の依頼した依頼の詳細を取得します。  
        matchedSupporterId がある場合は、サポーター情報も返します。
      security:
        - bearerAuth: []
      parameters:
        - in: path
          name: requestId
          required: true
          schema:
            type: integer
            example: 12
      responses:
        "200":
          description: 依頼詳細取得成功
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/RequestDetailUser"
        "400":
          description: リクエストIDは数値を入力してください。
        "403":
          description: あなたの依頼ではありません。
        "404":
          description: 依頼が見つかりません。
        "500":
          description: サーバーエラー
  /supporter/requests:
    get:
      tags:
        - Supporter
      summary: サポーター向け募集中依頼一覧を取得
      description: |
        ログイン中のサポーター向けに「open」の依頼一覧を返します。  
        ただし、そのサポーターが decline(辞退した） / refusal（交代要請された）依頼は除外されます。
      security:
        - bearerAuth: []
      responses:
        "200":
          description: 依頼一覧取得成功
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/SupporterRequest"
        "403":
          description: サポーターのみ閲覧可能です。
        "500":
          description: サーバーエラー
  /supporter/requests/{requestId}:
    get:
      tags:
        - Supporter
      summary: サポーター向けの依頼詳細を取得
      description: |
        依頼の詳細情報を返します。  
        - 募集中(open)の場合：利用者情報は簡易表示  
        - 自分が引き受けた依頼：利用者の詳細情報を返す  
        - 他のサポーターが引き受けている場合(異常系)：利用者情報は簡易表示
      security:
        - bearerAuth: []
      parameters:
        - in: path
          name: requestId
          required: true
          schema:
            type: integer
            example: 12
      responses:
        "200":
          description: 依頼詳細取得成功
          content:
            application/json:
              schema:
                oneOf:
                  - $ref: "#/components/schemas/SupporterRequestDetailOpen"
                  - $ref: "#/components/schemas/SupporterRequestDetailMatched"

              examples:
                openCase:
                  summary: open の場合のuser情報
                  value:
                    id:
                      type: integer
                    title:
                      type: string
                    description:
                      type: string
                    status:
                      type: string
                      enum: [open, matched, confirmed, completed, canceled, expired]
                      example: "open"
                    scheduledDate:
                      type: string
                      format: date
                    scheduledStartTime:
                      type: string
                    scheduledEndTime:
                      type: string
                    scheduledDurationMinutes:
                      type: integer
                      nullable: true
                    workLocation1:
                      type: string
                    workLocation2:
                      type: string
                    matchedSupporterId:
                      type: string
                      nullable: true
                    centerId:
                      type: string
                    createdAt:
                      type: string
                      format: date-time
                    updatedAt:
                      type: string
                      format: date-time
                    user:
                      ageGroup: "80代"
                      gender: "女性"
                      address1: "旭町1丁目"
                      bio: "難聴があり聞き取りにくいので、ゆっくり話してください。"

                matchedCase:
                  summary: matched の場合のuser情報
                  value:
                    id:
                      type: integer
                    title:
                      type: string
                    description:
                      type: string
                    status:
                      type: string
                      enum: [open, matched, confirmed, completed, canceled, expired]
                      example: "open"
                    scheduledDate:
                      type: string
                      format: date
                    scheduledStartTime:
                      type: string
                    scheduledEndTime:
                      type: string
                    scheduledDurationMinutes:
                      type: integer
                      nullable: true
                    workLocation1:
                      type: string
                    workLocation2:
                      type: string
                    matchedSupporterId:
                      type: string
                      nullable: true
                    centerId:
                      type: string
                    createdAt:
                      type: string
                      format: date-time
                    updatedAt:
                      type: string
                      format: date-time
                    user:
                      familyName: "井上"
                      firstName: "節子"
                      familyNameKana: "イノウエ"
                      firstNameKana: "セツコ"
                      gender: "女性"
                      phoneNumber: "090-1111-1111"
                      address1: "旭町1丁目"
                      address2: "42-55"
                      profileImageUrl: "/sampleuser6.png"
                      bio: "難聴があり聞き取りにくいので、ゆっくり話してください。"
                      ageGroup: "80代"
        "400":
          description: リクエストIDは数値を入力してください。
        "403":
          description: |
           閲覧権限なし  
           - サポーターのみ閲覧可能です。 
           - この依頼は他のサポーターが対応中のため閲覧できません。
        "404":
          description: |
           該当データなし
           - 依頼が見つかりません。
           - 利用者情報が見つかりません。
        "500":
          description: サーバーエラー
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
...
    UserRequest:
      type: object
      properties:
        id:
          type: integer
          example: 123
        title:
          type: string
          example: "買い物代行"
        description:
          type: string
          example: "鼻炎の薬をお願いします"
        status:
          type: string
          enum: [open, matched, confirmed, completed, canceled, expired]
          example: "open"
        scheduledDate:
          type: string
          format: date
          example: "2025-11-10"
        scheduledStartTime:
          type: string
          example: "10:00"
        scheduledEndTime:
          type: string
          example: "11:00"
        scheduledDurationMinutes:
          type: integer
          nullable: true
        workLocation1:
          type: string
          example: "マツモトキヨシ"
        workLocation2:
          type: string
          example: "旭町1丁目"
        matchedSupporterId:
          type: string
          nullable: true
        centerId:
          type: string
          example: "C001"
        createdAt:
          type: string
          format: date-time
          example: "2025-11-01T10:15:30Z"
        updatedAt:
          type: string
          format: date-time
          example: "2025-11-01T10:15:30Z"
        
    # =====================================================
    #  新規依頼(入力) NewRequestInpit
    # =====================================================
    NewRequestInput:
      type: object
      properties:
        description:
          type: string
          example: "鼻炎の薬をお願いします"
        scheduledDate:
          type: string
          format: date
          example: "2025-11-10"
        scheduledStartTime:
          type: string
          example: "10:00"
        scheduledEndTime:
          type: string
          example: "11:00"
        workLocation1:
          type: string
          example: "マツモトキヨシ"
    
    # =====================================================
    #  新規依頼(出力) NewRequest
    # =====================================================
    NewRequest:
      type: object
      properties:
        title:
          type: string
          example: "買い物代行"
        description:
          type: string
          example: "鼻炎の薬をお願いします"
        scheduledDate:
          type: string
          format: date
          example: "2025-11-10"
        scheduledStartTime:
          type: string
          example: "10:00"
        scheduledEndTime:
          type: string
          example: "11:00"
        workLocation1:
          type: string
          example: "マツモトキヨシ"
        workLocation2:
          type: string
          example: "旭町1丁目"
        cenerId:
          type: string
          example: "C001"

    # =====================================================
    #  利用者向け詳細 RequestDetailUser
    # =====================================================
    RequestDetailUser:
      type: object
      properties:
        id:
          type: integer
          example: 123
        title:
          type: string
          example: "買い物代行"
        description:
          type: string
          example: "鼻炎の薬をお願いします"
        status:
          type: string
          enum: [open, matched, confirmed, completed, canceled, expired]
          example: "matched"
        scheduledDate:
          type: string
          format: date
          example: "2025-11-10"
        scheduledStartTime:
          type: string
          example: "10:00"
        scheduledEndTime:
          type: string
          example: "11:00"
        scheduledDurationMinutes:
          type: integer
          nullable: true
        workLocation1:
          type: string
          example: "マツモトキヨシ"
        workLocation2:
          type: string
          example: "旭町1丁目"
        matchedSupporterId:
          type: string
          example: "BuUmfPzwDpfZi4mtPACvdQrMiE73"
        centerId:
          type: string
          example: "C001"
        createdAt:
          type: string
          format: date-time
          example: "2025-11-01T10:15:30Z"
        updatedAt:
          type: string
          format: date-time
          example: "2025-11-01T10:15:30Z"
        supporter:
          type: object
          nullable: true
          description: matchedSupporterId がある場合のみ表示
          properties:
            supporterName:
              type: string
              example: "木下亜里沙"
            supporterPhone:
              type: string
              example: "080-3333-3333"
            supporterNote:
              type: string
              example: "お役に立てると嬉しいです！"
            supporterAvatarUrl:
              type: string
              example: "/sampleuser3.png"

    # =====================================================
    #  サポーター向け募集中依頼一覧 SupporterRequest
    # =====================================================
    SupporterRequest:
      type: object
      properties:
        id:
          type: integer
        title:
          type: string
        description:
          type: string
        status:
          type: string
          enum: [open, matched, confirmed, completed, canceled, expired]
          example: "open"
        scheduledDate:
          type: string
          format: date
        scheduledStartTime:
          type: string
        scheduledEndTime:
          type: string
        scheduledDurationMinutes:
          type: integer
          nullable: true
        workLocation1:
          type: string
        workLocation2:
          type: string
        matchedSupporterId:
          type: string
          nullable: true
        centerId:
          type: string
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
        user:
          type: object
          properties:
            id:
              type: string
            gender:
              type: string
            address1:
              type: string
            ageGroup:
              type: string
              example: "80代"
            bio:
              type: string

    # =====================================================
    # サポーター向け詳細(引受前) SupporterRequestDetailOpen
    # =====================================================
    SupporterRequestDetailOpen:
      type: object
      properties:
        id:
          type: integer
        title:
          type: string
        description:
          type: string
        status:
          type: string
          enum: [open, matched, confirmed, completed, canceled, expired]
          example: "open"
        scheduledDate:
          type: string
          format: date
        scheduledStartTime:
          type: string
        scheduledEndTime:
          type: string
        scheduledDurationMinutes:
          type: integer
          nullable: true
        workLocation1:
          type: string
        workLocation2:
          type: string
        matchedSupporterId:
          type: string
          nullable: true
        centerId:
          type: string
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
        user:
          type: object
          properties:
            ageGroup:
              type: string
              example: "80代"
            gender:
              type: string
            address1:
              type: string
            bio:
              type: string

    # =====================================================
    # サポーター向け詳細(引受済) SupporterRequestDetailMatched
    # =====================================================
    SupporterRequestDetailMatched:
      type: object
      properties:
        id:
          type: integer
        title:
          type: string
        description:
          type: string
        status:
          type: string
          enum: [open, matched, confirmed, completed, canceled, expired]
          example: "matched"
        scheduledDate:
          type: string
          format: date
        scheduledStartTime:
          type: string
        scheduledEndTime:
          type: string
        scheduledDurationMinutes:
          type: integer
          nullable: true
        workLocation1:
          type: string
        workLocation2:
          type: string
        matchedSupporterId:
          type: string
          nullable: true
        centerId:
          type: string
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
        user:
          type: object
          properties:
            familyName:
              type: string
            firstName:
              type: string
            familyNameKana:
              type: string
            firstNameKana:
              type: string
            gender:
              type: string
            phoneNumber:
              type: string
            address1:
              type: string
            address2:
              type: string
            profileImageUrl:
              type: string
            bio:
              type: string
            ageGroup:
              type: string
              example: "80代"

    # =====================================================
    # サポーター向け引受一覧 SupporterOrderSummary
    # =====================================================
...